djangoSamples/pizzaOrder/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-crispy-forms==2.3
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
urllib3==2.2.3
